Plugins: Fixed compilation error on gcc
Former-commit-id: 6ce3f81b00d4f3d88ffe9a7bbb8962e3a9437f56

+++ plugins/processing/classification/src/box-algorithms/ovpCBoxAlgorithmClassifierTrainer.cpp
@@ -438,7 +438,7 @@ float64 CBoxAlgorithmClassifierTrainer::getAccuracy(const size_t uiStartIndex, c
 
 	if(uiStopIndex-uiStartIndex==0)
 	{
-		this->getLogManager() << LogLevel_Error << "Start and stop indexes are the same (" << uiStartIndex << ")\n";
+		this->getLogManager() << LogLevel_Error << "Start and stop indexes are the same (" << static_cast<uint64>(uiStartIndex) << ")\n";
 		return 0;
 	}
